+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Release
 
 
+- Added the ModifiedZScore Class (22/06/2022)
 - Added the ZScore Class (22/06/2022)
 - Added the outliers file (22/06/2022)
 - Added the Dixon Module (13/06/2022)