+++ docs/changelog.md
@@ -7,6 +7,7 @@
 ### Release
 
 
+- Added documentation for the ModifiedZScore test (22/06/2022)
 - Added documentation for the ZScore test (22/06/2022)
 - Added documentation for the Dixon Module (13/06/2022)